shared: Move PollData into shared.
